Implement rotation sleeping
This fixes an issue where CFrame springs with rotation but no translation terminated too early and snapped to the end value.

This isn't perfect - There are preexisting numerical precision issues with the exponential and logarithmic maps that can prevent CFrame springs from ever terminating. This can lead to performance problems.

For now, we will take the stance that correctness trumps performance, and ignore it. :D