Add completion callbacks
This change adds spr.completed, which registers a function to be called the next time an instance stops animating.

It is a bad idea in general to link business logic to animation states, but knowing when an instance has been let go by spr is useful for tracking object lifetime (e.g. making an object invisible and then destroying it).