Fix camera jitter
## Problem
Using spr.target to follow a simulated part with a camera produces noticeable jitter. This is because there is an ordering conflict between the user calling spr.target in a heartbeat loop, while we update and set the properties on another heartbeat loop.

## Solution
Take advantage of the task scheduler and force the ordering to be correct. The task order must be

1. spr.target instances that are not cameras in spr
1. Update instances that are not cameras in spr
2. spr.target cameras
3. Update cameras in spr

Cameras should now be updated on Heartbeat:

```lua
RunService.Heartbeat:Connect(function()
    spr.target(workspace.CurrentCamera, 1, 2, {
        CFrame = target.CFrame * targetToCamera
    })
end)
```

If you want your camera to follow a part that is in turn being controlled by spr:
- spr.target the part on PreAnimation, &&
- spr.target the camera on Heartbeat

```lua
RunService.PreAnimation:Connect(function()
    spr.target(target, 1, 2, {
        CFrame = getTargetCFrame()
    })
end)

RunService.Heartbeat:Connect(function()
    spr.target(workspace.CurrentCamera, 1, 2, {
        CFrame = target.CFrame * targetToCamera
    })
end)
```